Second step of business logic refactoring
